+++ src/main/java/com/centum/repository/AlbumRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface AlbumRepository extends JpaRepository<Album, Long> {
-    List<Album> findByArtistId(Long artistId);
+public interface AlbumRepository extends JpaRepository<Album, Integer> {
+    List<Album> findByArtistId(Integer artistId);
 
     Optional<Album> findBySpotifyId(String spotifyId);
 }
