+++ src/main/java/com/centum/repository/ArtistRepository.java
@@ -4,6 +4,6 @@ import com.centum.entity.Artist;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface ArtistRepository extends JpaRepository<Artist, Long> {
+public interface ArtistRepository extends JpaRepository<Artist, Integer> {
     Optional<Artist> findBySlug(String slug);
 }
