+++ src/main/java/com/centum/repository/FavoriteRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface FavoriteRepository extends JpaRepository<Favorite, Long> {
-    List<Favorite> findByUsuarioId(Long usuarioId);
+public interface FavoriteRepository extends JpaRepository<Favorite, Integer> {
+    List<Favorite> findByUsuarioId(Integer usuarioId);
 
-    Optional<Favorite> findByUsuarioIdAndAlbumId(Long usuarioId, Long albumId);
+    Optional<Favorite> findByUsuarioIdAndAlbumId(Integer usuarioId, Integer albumId);
 }
