+++ src/main/java/com/centum/repository/WishlistRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface WishlistRepository extends JpaRepository<Wishlist, Long> {
-    List<Wishlist> findByUsuarioId(Long usuarioId);
+public interface WishlistRepository extends JpaRepository<Wishlist, Integer> {
+    List<Wishlist> findByUsuarioId(Integer usuarioId);
 
-    Optional<Wishlist> findByUsuarioIdAndAlbumId(Long usuarioId, Long albumId);
+    Optional<Wishlist> findByUsuarioIdAndAlbumId(Integer usuarioId, Integer albumId);
 }
